+++ frontend/src/views/scrum/sprints/Sprints.vue
@@ -76,6 +76,16 @@
                 Diagrama de quemado (Burn-down/up)
               </span>
             </v-tooltip>
+            <v-tooltip bottom>
+              <template #activator="{ on, attrs }">
+                <v-btn icon :to="{ name: 'sprint-gantt', params: { sprintId: item.id } }" v-bind="attrs" v-on="on">
+                  <v-icon>mdi-chart-timeline</v-icon>
+                </v-btn>
+              </template>
+              <span>
+                Diagrama de Gantt
+              </span>
+            </v-tooltip>
           </template>
         </ItemIndex>
       </v-col>